feat: Write brute force version of `all_wordifications()`
Just doing the brute force version first as a starting point and refining it later.

The function first converts all numbers to a list of possible letters and then iterates through all permutations of those letters. For each permutation, it checks to see if any substrings (starting from the back) are English words. If so, it concatenates the number and word and adds it to the list.

The function does use PyEnchant for the English dictionary which is not a standard Python library but since I don't think the purpose of this exercise is to write an English dictionary lookup library I think it's probably okay.

I also wrote `words_to_number()` but that was just a simple lookup table essentially.