feat: Implement `number_to_words()` using `all_wordifications()`
For a quick implementation, I found all possible wordifications from a US phone number and then simply selected one at random from the list and returned it.

Since the exercise is a little open-ended regarding what the expected output should be, I think the next step would be to return the phone number with the longest word in it. Beyond that, I think the main challenge that needs to be tackled before it can be considered complete is improving the efficiency of `all_wordifications()` and possibly loosening the constraints to not just find words at the end of the phone number.