feat: Improve `all_wordifications()`
This commit adds the ability to find wordifications that don't just have the word on the end of the number so "1-800-PAINTER" and "1-800-PAINT-37" are both valid for "1-800-724-6837". This adds computational overhead to the algorithm but since all US phone numbers are 10 digits long it's still constant time.

It also adds tests for the necessary utility functions and a simple script for easily running all tests from the command line.